requirements.txt
tensorflow
keras
mtcnn
Pillow
opencv-python
streamlit